memory: Move to common
The memory allocation algorithm can be used for kernel and user processes
alike. To avoid code duplication, we move the code to common to reuse
it in libcedos.

To this end, we make the start and end addresses of the allocatable
memory region flexible and set them as arguments in malloc_init.