kernel: Add alarms
Alarms implement the sleep function on a low level. The alarm_tick function
is called at every system tick and decreases all existing alarms. If one
alarm reaches 0, an action is taken depending on the alarm mode; Either
the program is woken up (in case of sleep), or the program is killed.

A new syscall is introduced: Sleep blocks the process for a given
number of ticks.