sched: Use kernel entry as idle process
Previously, we created a new page directory for the entry process and had a whole lot of if statements in the scheduler to treat the idle process specially.

Instead, we now just use the page directory we got from the bootloader, and enter the idle process from the kernel main. That way, we don't waste pages, we keep the startup context on the idle process stack, and have a cleared-up scheduler function that does not occlude what happens in PID 0.